Move struct TowerReferencePoint  class in reference_point_elast.h
