+++ include/seahowl/elasto/elasto.h
@@ -12,10 +12,15 @@ class ChElementBeamTaperedTimoshenko;
 }  // namespace fea
 }  // namespace chrono
 
+#include <seahowl/elasto/reference_point_elasto.h>
+#include <seahowl/elasto/utils_elasto.h>
+
 
 namespace seahowl {
 ///@brief Elastodynamic model module
 namespace elasto {
+
+struct BladeReferencePointElasto;
 
 /**@biref Component <<interface>>
 
@@ -27,6 +32,9 @@ class ElastoComponent {
     virtual void rotate(double angle, chrono::ChVector<double> axis) const = 0; ///< Rotate the system
     virtual void translate(chrono::ChVector<double> translation_vector) const = 0; ///< Translate the system
     virtual double get_mass() const = 0; ///< Get total mass
+    std::vector<double> discretization_fractions; ///< Fraction ?
+    std::vector<BladeReferencePointElasto> reference_points; ///< Original points
+    std::vector<BladeReferencePointElasto> discretized_points; ///< Discretized point
     ///@}
 };
 
@@ -40,7 +48,9 @@ class ElastoFEAComponent : public ElastoComponent {
     virtual void rotate(double angle, chrono::ChVector<double> axis) const override;     ///< Rotate the system
     virtual void translate(chrono::ChVector<double> translation_vector) const override;  ///< Translate the system
     virtual double get_mass() const override;                                            ///< Get total mass
+    virtual void set_damping_coefficients(double axial, double edge, double flap, double torsion) = 0;
     ///@}
 };
+
 }  // namespace elasto
 }  // namespace seahowl
