+++ include/seahowl/elasto/reference_point_elasto.h
@@ -77,5 +77,62 @@ struct BladeReferencePointElasto {
     };
 };
 
+
+
+
+
+
+/**@brief Tower (DOF) reference point */
+struct TowerReferencePoint {
+    chrono::ChVector<double> coordinates;
+    double fraction;
+    double density;
+    double stiffness_axial;
+    double stiffness_foreaft;
+    double stiffness_sideside;
+    double stiffness_torsion;
+    chrono::fea::DampingCoefficients damping_coefficients;
+
+    TowerReferencePoint operator*(const double factor) const {
+        TowerReferencePoint new_point;
+        new_point.coordinates = coordinates * factor;
+        new_point.fraction = fraction * factor;
+        new_point.density = density * factor;
+        new_point.stiffness_axial = stiffness_axial * factor;
+        new_point.stiffness_foreaft = stiffness_foreaft * factor;
+        new_point.stiffness_sideside = stiffness_sideside * factor;
+        new_point.stiffness_torsion = stiffness_torsion * factor;
+        new_point.damping_coefficients.bx = damping_coefficients.bx * factor;
+        new_point.damping_coefficients.bx = damping_coefficients.bx * factor;
+        new_point.damping_coefficients.bx = damping_coefficients.bx * factor;
+        new_point.damping_coefficients.bx = damping_coefficients.bx * factor;
+        new_point.damping_coefficients.by = damping_coefficients.by * factor;
+        new_point.damping_coefficients.bz = damping_coefficients.bz * factor;
+        new_point.damping_coefficients.bt = damping_coefficients.bt * factor;
+        new_point.damping_coefficients.alpha = damping_coefficients.alpha * factor;
+        return new_point;
+    };
+
+    TowerReferencePoint operator+(const TowerReferencePoint& other) const {
+        TowerReferencePoint new_point = *this;
+        new_point.coordinates += other.coordinates;
+        new_point.fraction += other.fraction;
+        new_point.density += other.density;
+        new_point.stiffness_axial += other.stiffness_axial;
+        new_point.stiffness_foreaft += other.stiffness_foreaft;
+        new_point.stiffness_sideside += other.stiffness_sideside;
+        new_point.stiffness_torsion += other.stiffness_torsion;
+        new_point.damping_coefficients.bx += other.damping_coefficients.bx;
+        new_point.damping_coefficients.bx += other.damping_coefficients.bx;
+        new_point.damping_coefficients.bx += other.damping_coefficients.bx;
+        new_point.damping_coefficients.bx += other.damping_coefficients.bx;
+        new_point.damping_coefficients.by += other.damping_coefficients.by;
+        new_point.damping_coefficients.bz += other.damping_coefficients.bz;
+        new_point.damping_coefficients.bt += other.damping_coefficients.bt;
+        new_point.damping_coefficients.alpha += other.damping_coefficients.alpha;
+        return new_point;
+    };
+};
+
 }  // namespace elasto
 }  // namespace seahowl
